requirements.txt
fastapi==0.111.1
langchain==0.2.12
python-dotenv==1.0.0
pydantic==1.10.8
langchain-community==0.2.11
langchain-core==0.2.29
langchain-groq==0.1.9
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
SQLAlchemy==2.0.32
python-jose==3.3.0
passlib==1.7.4
psycopg2==2.9.9
bcrypt==4.2.0